fixed timer

+++ src/util/chk_time.h
@@ -110,7 +110,7 @@ static inline uint64_t chk_systick64() {
     return c->last_time;
 }
 
-static inline uint32_t chk_accurate_tick64(){
+static inline uint64_t chk_accurate_tick64(){
     return _clock_time();
 }
